+++ scripts/crafting/harvesting.js
@@ -35,6 +35,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Carrion Crawler",
+        "dc": 12,
         "rewards": [
             {
                 "name": "Carrion Crawler Mucus"
@@ -67,6 +68,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Ettercap",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Cap of Ettercap Fungi"
@@ -96,6 +98,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Giant Wasp",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Wasp Venom"
@@ -120,6 +123,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Myconid Sovereign",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Cap of Myconid Fungi"               
@@ -128,6 +132,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Phase Spider",
+        "dc": 12,
         "rewards": [
             {
                 "name": "Phase Spider Venom"              
@@ -136,6 +141,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Poisonus Snake",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Snake Venom"
@@ -144,6 +150,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Poisonus Spider",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Spider Venom"
@@ -152,6 +159,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Purple Worm",
+        "dc": 18,
         "rewards": [
             {
                 "name": "Purple Worm Venom"
@@ -173,6 +181,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Scorpion",
+        "dc": 11,
         "rewards": [
             {
                 "name": "Scorpion Venom"
@@ -205,6 +214,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Young Green Dragon",
+        "dc": 14,
         "rewards": [
             {
                 "name": "Young Green Dragon Venom"
@@ -213,6 +223,7 @@ const HARVESTABLES = [
     },
     {
         "name": "Wyvern",
+        "dc": 13,
         "rewards": [
             {
                 "name": "Wyvern Venom"